+++ certbot-dns-cloudflare/docs/api/dns_cloudflare.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_cloudflare.dns_cloudflare`
+:mod:`certbot_dns_cloudflare._internal.dns_cloudflare`
 --------------------------------------------
 
-.. automodule:: certbot_dns_cloudflare.dns_cloudflare
+.. automodule:: certbot_dns_cloudflare._internal.dns_cloudflare
    :members:
